* Fixed a couple missing lines. * Removed webpack. (part of cross-project removing of webpack from stuff that doesn't need it -- possibly everything other than websites, but at least all libs [reason might still want webpack for non-website [non-lib] projects, is that it lets you do npm package-patching and misc. stuff like that, whereas for libs that's rarely needed and adds complications for things like source-maps]) [self: see notes for more info]
Note: For some reason, the ".js.map" files output in the Dist folder are not loaded by webpack (at least in the CD project). However, I think that's a misconfiguration in the CD webpack, rather than this lib's ts-config. (so will leave the source-map outputting on for now -- nothing's currently using it, but it's pretty cheap as well since they're separate files, so don't have much incrementing storage cost)